Added Scenario ObjGen, SkyBox ObjGen and organized Data
Added Scenario ObjGen akin to MapZone/NodeScene with it's own tracker, start, delay, interval, event, interaction and exit sections. Can configure as site default and set a MapZone to start in as well as the scenario's title and instructions which are updated in HTML. Added SkyBox ObjGen for Lights, Sky & Space items. Organized Data and Objects.